requirements.txt
flask
google-genai
pandas
numpy
scikit-learn
python-dotenv